update for latest api
